Moving the rest of the windows/views into Views/
