+++ DW3ArduinoWinDev/DW3ArduinoEditor/ViewModels/MainWindowViewModel.cs
@@ -3,6 +3,7 @@ using DW3ArduinoEditor.Enums;
 using DW3ArduinoEditor.Graphics;
 using DW3ArduinoEditor.SaveData;
 using DW3ArduinoEditor.Types;
+using DW3ArduinoEditor.Views;
 using System.Collections.ObjectModel;
 using System.IO;
 using System.Text.Json;
