+++ DW3ArduinoWinDev/DW3ArduinoEditor/Views/ResizeTileMapWindow.xaml
@@ -1,4 +1,4 @@
-﻿<Window x:Class="DW3ArduinoEditor.ResizeTileMapWindow"
+﻿<Window x:Class="DW3ArduinoEditor.Views.ResizeTileMapWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
